requirements.txt
asgiref
binaryornot==0.4.4
certifi==2026.1.4
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cookiecutter==2.6.0
Django==6.0.2
django-formtools==2.5.1
idna==3.11
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
psycopg2-binary==2.9.11
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-slugify==8.0.4
PyYAML==6.0.3
requests==2.32.5
rich==14.3.1
six==1.17.0
sqlparse==0.5.5
text-unidecode==1.3
tzdata==2025.3
urllib3==2.6.3
djangorestframework==3.16.1
drf-yasg==1.21.11
django-ckeditor==6.7.3
django-cors-headers
jdatetime
setuptools
arrow
num2words
django-mptt>=0.14.0
django-hordak==2.0.0a2
weasyprint
gunicorn
